feat: very adjusts

+++ src/repository/userRepository.ts
@@ -1,5 +1,9 @@
-import { ICreateUserBot, IUserBotConfigCreate, IUserBotConfigUpdate } from '@/interfaces/UserBot';
-import { PrismaClient } from '@prisma/client';
+import {
+  ICreateUserBot,
+  IUserBotConfigUpdate,
+  IUserSettingsTrading,
+} from "@/interfaces/UserBot";
+import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient();
 
@@ -12,18 +16,56 @@ export class UserBotConfigRepository {
     return prisma.userBotConfig.findMany();
   }
 
-  async findById(id: number) {
-    return prisma.userBotConfig.findUnique({ where: { id } });
+  async findTradingSettingsByUserId(
+    userId: number
+  ): Promise<IUserSettingsTrading | null> {
+    return prisma.userBotConfig.findUnique({
+      select: {
+        key: true,
+        secret: true,
+        passphrase: true,
+        riskThreshold: true,
+        amountForSetMargin: true,
+        quantity: true,
+        variation: true,
+        leverage: true,
+        balance: true,
+        profitPercentage: true,
+        from: true,
+        evenPositive: true,
+        evenNegative: true,
+        accountBalance: true,
+        availableAccountBalance: true,
+      },
+      where: {
+        id: userId,
+      },
+    });
   }
 
-  async update(id: number, data: IUserBotConfigUpdate) {
+  async findByUserId(userId: number) {
+    return prisma.userBotConfig.findUnique({ where: { id: userId } });
+  }
+
+  async update(userId: number, data: IUserBotConfigUpdate) {
     return prisma.userBotConfig.update({
-      where: { id },
+      where: { id: userId },
       data,
     });
   }
 
-  async delete(id: number) {
-    return prisma.userBotConfig.delete({ where: { id } });
+  async delete(userId: number) {
+    return prisma.userBotConfig.delete({ where: { id: userId } });
+  }
+
+  async updateAvailableAccountBalance(value: number, userId: number) {
+    return prisma.userBotConfig.update({
+      data: {
+        availableAccountBalance: value,
+      },
+      where: {
+        id: userId,
+      },
+    });
   }
 }
